// src/client/types.gen.ts
export type CreatePostDto = {
  /**
   * 文章标题
   */
  title: string
  /**
   * 文章摘要
   */
  excerpt?: string
  /**
   * 文章内容
   */
  content: string
  /**
   * 分类 ID
   */
  category_id?: number
  /**
   * 标签名称列表
   */
  tag_names?: Array<string>
}

export type UpdatePostDto = {
  /**
   * 文章标题
   */
  title?: string
  /**
   * 文章摘要
   */
  excerpt?: string
  /**
   * 文章内容
   */
  content?: string
  /**
   * 分类 ID
   */
  category_id?: number
  /**
   * 标签名称列表
   */
  tag_names?: Array<string>
}